event/requirements.txt
paho-mqtt
dynaconf
requests
